[SUS PUSH] Waiting before ledge jumping attempt
WARNING: MAY BE BAD (HALF IMPLEMENTED FEATURED BUT IM PUSHING ANYWAY)
I'm trying to make it so that the player must wait before jumping from a ledge after landing on it. I realize my code is fucked up and messy as hell. I'll explain what I can, but basically I am struggling with the logic portion of getting this to work.